+++ doc/src/changelog.md
@@ -6,6 +6,8 @@ Version History
 -   RTCIntersectContext always non-const in Embree API
 -   Made RTCHit aligned to 16 bytes in Embree API
 -   Bugfix in AVX512VL codepath for rtcIntersect1
+-   For sphere geometries the intersection filter gets now invoked for
+    front and back hit.
 
 ### New Features in Embree 3.7.0
 -   Added quaternion motion blur for correct interpolation of rotational transformations.